Move faer to optional
